Add files via upload
Wrote the code for airline analysis: Used the flights data set along with the airlines data set to create bar charts that showed the overall average delay (departure delay + arrival delay) of each airline. For further analysis I also created the code and bar chart to look at overall delays that were filtered to over an hour. I found within the datasets that certain airlines had more flights and so this would skew the averages for the first plot. Filtering the delay to more than an hour fine tuned how punctual each airline carrier was and gave a more definitive answer to which airline came out as most punctual. 
Wrote the code for the interactive airport delays map and top 6 airports table: Used the airports dataset and joined it to the flights dataset to use the latitude and longitude information to plot it onto a map of the United States. I used a bubble plot on the map that shows the average total delays- the larger the dot the more the delay. I labelled each state and used the plotly function so that the user could hover over each dot to show the airport, lat and long and average delay in minutes. To gain more clarity into which airport was most punctual I created the code that looks at the top 6 airports in ascending order (this is shown in our presentation). 
Organised the pdf pages and made graphs side by side or widened to fit into four pages: In our original HTML markdown it was difficult to see how many pages we had and whether we could fit everything in. I edited the original file into a pdf file and changed the code in the chunks to make sure all the graphs would fit whilst also be readable and clear.